// project-management-app/src/app/models/Filter.ts
export class Filter {

    username: string;
    firstName: string;
    lastName: string;
    legalIdNumber: string;
    apogeeCode: string;
    studentNationalCode: string;
    profileCode: string;
    sortBy: string;
    sortDirection: string;
    page: number;
    pageSize: number;
    count: number;
    countOnly: boolean;
}